Change to Long

+++ src/main/java/org/kpmp/participant/ParticipantSummaryDatasetRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParticipantSummaryDatasetRepository extends CrudRepository<ParticipantSummaryDataset, Integer> {
+public interface ParticipantSummaryDatasetRepository extends CrudRepository<ParticipantSummaryDataset, Long> {
 
 	@Cacheable("participantByRedcap")
 	ParticipantSummaryDataset findByRedcapId(@Param("redcap_id") String redcapId);
